refactor(system): rename discovery models module to keys
The module defines ModelKey and TagProviderKey -- key types, not domain
models. Rename clarifies intent and avoids confusion with theme model
modules elsewhere in the codebase.
