+++ packages/overture-schema-system/src/overture/schema/system/discovery/__init__.py
@@ -5,7 +5,7 @@ from .discovery import (
     filter_models,
     get_registered_model,
 )
-from .models import ModelKey
+from .keys import ModelKey
 from .types import ModelDict
 
 __all__ = [
